requirements.txt
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
psycopg==3.1.18
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
typing_extensions==4.9.0